move basic data type classes on japan module
