+++ lib/zakuro/era/japan/gengou/resource/type.rb
@@ -4,6 +4,9 @@ require_relative '../../../western/calendar'
 
 require_relative '../../calendar'
 
+require_relative '../../type/base/switch_date'
+require_relative '../../type/base/both/year'
+
 # :nodoc:
 module Zakuro
   # :nodoc:
@@ -20,9 +23,9 @@ module Zakuro
         class Gengou
           # @return [String] 元号名
           attr_reader :name
-          # @return [Both::Year] 開始年（和暦/西暦）
+          # @return [Type::Base::Both::Year] 開始年（和暦/西暦）
           attr_reader :start_year
-          # @return [SwitchDate] 開始日（和暦/西暦）
+          # @return [Type::Base::SwitchDate] 開始日（和暦/西暦）
           attr_reader :start_date
           # @return [Integer] 終了年
           attr_reader :last_year
@@ -33,15 +36,15 @@ module Zakuro
           # 初期化
           #
           # @param [String] name 元号名
-          # @param [Both::Year] start_year 開始年（和暦/西暦）
-          # @param [SwitchDate] start_date 開始日（和暦/西暦）
+          # @param [Type::Base::Both::Year] start_year 開始年（和暦/西暦）
+          # @param [Type::Base::SwitchDate] start_date 開始日（和暦/西暦）
           # @param [Integer] last_date 終了年
           # @param [Western::Calendar] last_date 終了日
           #
-          def initialize(name: '', start_year: Both::Year.new,
-                         start_date: SwitchDate.new,
+          def initialize(name: '', start_year: Type::Base::Both::Year.new,
+                         start_date: Type::Base::SwitchDate.new,
                          last_date: Western::Calendar.new,
-                         last_year: Both::Year::INVALID)
+                         last_year: Type::Base::Both::Year::INVALID)
             @name = name
             @start_year = start_year
             @start_date = start_date
@@ -185,9 +188,9 @@ module Zakuro
           attr_reader :id
           # @return [String] 元号セット名
           attr_reader :name
-          # @return [Both::Year] 元号セットでの終了年
+          # @return [Type::Base::Both::Year] 元号セットでの終了年
           attr_reader :last_year
-          # @return [Both::Date] 元号セットでの終了日
+          # @return [Type::Base::Both::Date] 元号セットでの終了日
           attr_reader :last_date
           # @return [Array<Gengou>] 元号リスト
           attr_reader :list
@@ -200,8 +203,8 @@ module Zakuro
           # @param [Western::Calendar] last_date 元号セットでの終了日
           # @param [Array<Gengou>] list 元号リスト
           #
-          def initialize(id: INVALID, name: '', last_year: Both::Year.new,
-                         last_date: Both::Date.new, list: [])
+          def initialize(id: INVALID, name: '', last_year: Type::Base::Both::Year.new,
+                         last_date: Type::Base::Both::Date.new, list: [])
             @id = id
             @name = name
             @last_year = last_year
@@ -232,131 +235,6 @@ module Zakuro
           #
           def invalid?
             @id == INVALID
-          end
-        end
-
-        #
-        # SwitchDate 切替日（運用/計算）
-        #
-        class SwitchDate
-          # @return [Both::Date] 計算値
-          attr_reader :calculation
-          # @return [Both::Date] 運用値
-          attr_reader :operation
-          # @return [True, False] 運用値
-          attr_reader :operated
-          # @return [Japan::Calendar] 和暦日
-          attr_reader :japan
-          # @return [Western::Calendar] 西暦日
-          attr_reader :western
-
-          #
-          # 初期化
-          #
-          # @param [Both::Date] calculation 計算値
-          # @param [Both::Date] operation 運用値
-          # @param [True, False] operated 運用値設定
-          #
-          def initialize(calculation: Both::Date.new, operation: Both::Date.new, operated: false)
-            @calculation = calculation
-            @operation = operation
-            @operated = operated
-
-            select
-          end
-
-          #
-          # 不正か
-          #
-          # @return [True] 不正
-          # @return [False] 不正なし
-          #
-          def invalid?
-            japan.invalid? || western.invalid?
-          end
-
-          private
-
-          def select
-            @japan = operation.japan
-            @western = operation.western
-
-            return if operated
-
-            calc_japan = calculation.japan
-            calc_western = calculation.western
-
-            @japan = calc_japan unless calc_japan.invalid?
-            @western = calc_western unless calc_western.invalid?
-          end
-        end
-
-        #
-        # Both 和暦/西暦
-        #
-        module Both
-          #
-          # Year 年
-          #
-          class Year
-            # @return [Integer] 不正値
-            INVALID = -1
-            # @return [Integer] 和暦元号年
-            attr_reader :japan
-            # @return [Integer] 西暦年
-            attr_reader :western
-
-            #
-            # 初期化
-            #
-            # @param [Integer] japan 和暦元号年
-            # @param [Integer] western 西暦年
-            #
-            def initialize(japan: INVALID, western: INVALID)
-              @japan = japan
-              @western = western
-            end
-
-            #
-            # 不正か
-            #
-            # @return [True] 不正
-            # @return [False] 不正なし
-            #
-            def invalid?
-              japan == INVALID || western == INVALID
-            end
-          end
-
-          #
-          # Date 日
-          #
-          class Date
-            # @return [Japan::Calendar] 和暦日
-            attr_reader :japan
-            # @return [Western::Calendar] 西暦日
-            attr_reader :western
-
-            #
-            # 初期化
-            #
-            # @param [Japan::Calendar] japan 和暦日
-            # @param [Western::Calendar] western 西暦日
-            #
-            def initialize(japan: Japan::Calendar.new, western: Western::Calendar.new)
-              @japan = japan
-              @western = western
-            end
-
-            #
-            # 不正か
-            #
-            # @return [True] 不正
-            # @return [False] 不正なし
-            #
-            def invalid?
-              japan.invalid? || western.invalid?
-            end
           end
         end
       end
