+++ spec/zakuro/calculation/era/gengou/internal/reserve/dated_list_spec.rb
@@ -241,13 +241,13 @@ describe 'Zakuro' do
                     Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
                       gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                         name: '元号1',
-                        start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                        start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                           japan: 1,
                           western: 450
                         ),
-                        start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                          calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                          operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                        start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                          calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                          operation: Zakuro::Japan::Type::Base::Both::Date.new(
                             japan: Zakuro::Japan::Calendar.parse(text: '元号1年1月1日'),
                             western: Zakuro::Western::Calendar.parse(text: '0450-01-12')
                           ),
@@ -285,13 +285,13 @@ describe 'Zakuro' do
                     Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
                       gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                         name: '元号1',
-                        start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                        start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                           japan: 1,
                           western: 450
                         ),
-                        start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                          calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                          operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                        start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                          calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                          operation: Zakuro::Japan::Type::Base::Both::Date.new(
                             japan: Zakuro::Japan::Calendar.parse(text: '元号1年1月1日'),
                             western: Zakuro::Western::Calendar.parse(text: '0450-01-01')
                           ),
@@ -303,11 +303,11 @@ describe 'Zakuro' do
                     Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
                       gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                         name: '元号2',
-                        start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                        start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                           japan: 1,
                           western: 450
                         ),
-                        start_date: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                        start_date: Zakuro::Japan::Type::Base::Both::Date.new(
                           japan: Zakuro::Japan::Calendar.new(
                             gengou: '元号2', year: 1, leaped: false, month: 1, day: 1
                           ),
@@ -356,13 +356,13 @@ describe 'Zakuro' do
                     Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
                       gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                         name: '元号1',
-                        start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                        start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                           japan: 1,
                           western: 450
                         ),
-                        start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                          calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                          operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                        start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                          calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                          operation: Zakuro::Japan::Type::Base::Both::Date.new(
                             japan: Zakuro::Japan::Calendar.parse(text: '元号1年1月1日'),
                             western: Zakuro::Western::Calendar.parse(text: '0450-01-01')
                           ),
@@ -374,13 +374,13 @@ describe 'Zakuro' do
                     Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
                       gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                         name: '元号2',
-                        start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                        start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                           japan: 1,
                           western: 450
                         ),
-                        start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                          calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                          operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                        start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                          calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                          operation: Zakuro::Japan::Type::Base::Both::Date.new(
                             japan: Zakuro::Japan::Calendar.parse(text: '元号2年1月1日'),
                             western: Zakuro::Western::Calendar.parse(text: '0450-01-21')
                           ),
@@ -392,11 +392,11 @@ describe 'Zakuro' do
                     Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
                       gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                         name: '元号3',
-                        start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                        start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                           japan: 1,
                           western: 450
                         ),
-                        start_date: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                        start_date: Zakuro::Japan::Type::Base::Both::Date.new(
                           japan: Zakuro::Japan::Calendar.new(
                             gengou: '元号3', year: 1, leaped: false, month: 1, day: 1
                           ),
@@ -451,13 +451,13 @@ describe 'Zakuro' do
                     Zakuro::Japan::Gengou::Alignment::LinearGengou.new(
                       gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                         name: '元号1',
-                        start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                        start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                           japan: 1,
                           western: 450
                         ),
-                        start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                          calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                          operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                        start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                          calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                          operation: Zakuro::Japan::Type::Base::Both::Date.new(
                             japan: Zakuro::Japan::Calendar.parse(text: '元号1年1月1日'),
                             western: Zakuro::Western::Calendar.parse(text: '0450-01-12')
                           ),
