+++ spec/zakuro/calculation/era/gengou/internal/reserve/empty_link_spec.rb
@@ -34,12 +34,12 @@ describe 'Zakuro' do
                   Zakuro::Calculation::Gengou::Counter.new(
                     gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                       name: '元号甲',
-                      start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                      start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1, western: 446
                       ),
-                      start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                        calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                        operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                      start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                        calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                        operation: Zakuro::Japan::Type::Base::Both::Date.new(
                           japan: Zakuro::Japan::Calendar.parse(text: '元号甲1年1月1日'),
                           western: Zakuro::Western::Calendar.parse(text: '0446-02-12')
                         ),
@@ -90,12 +90,12 @@ describe 'Zakuro' do
                   Zakuro::Calculation::Gengou::Counter.new(
                     gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                       name: '元号甲',
-                      start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                      start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1, western: 446
                       ),
-                      start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                        calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                        operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                      start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                        calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                        operation: Zakuro::Japan::Type::Base::Both::Date.new(
                           japan: Zakuro::Japan::Calendar.parse(text: '元号甲1年1月1日'),
                           western: Zakuro::Western::Calendar.parse(text: '0446-02-12')
                         ),
@@ -110,12 +110,12 @@ describe 'Zakuro' do
                   Zakuro::Calculation::Gengou::Counter.new(
                     gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                       name: '元号乙',
-                      start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                      start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1, western: 446
                       ),
-                      start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                        calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                        operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                      start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                        calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                        operation: Zakuro::Japan::Type::Base::Both::Date.new(
                           japan: Zakuro::Japan::Calendar.parse(text: '元号甲1年3月1日'),
                           western: Zakuro::Western::Calendar.parse(text: '0446-04-12')
                         ),
@@ -130,12 +130,12 @@ describe 'Zakuro' do
                   Zakuro::Calculation::Gengou::Counter.new(
                     gengou: Zakuro::Japan::Gengou::Resource::Gengou.new(
                       name: '元号丙',
-                      start_year: Zakuro::Japan::Gengou::Resource::Both::Year.new(
+                      start_year: Zakuro::Japan::Type::Base::Both::Year.new(
                         japan: 1, western: 446
                       ),
-                      start_date: Zakuro::Japan::Gengou::Resource::SwitchDate.new(
-                        calculation: Zakuro::Japan::Gengou::Resource::Both::Date.new,
-                        operation: Zakuro::Japan::Gengou::Resource::Both::Date.new(
+                      start_date: Zakuro::Japan::Type::Base::SwitchDate.new(
+                        calculation: Zakuro::Japan::Type::Base::Both::Date.new,
+                        operation: Zakuro::Japan::Type::Base::Both::Date.new(
                           japan: Zakuro::Japan::Calendar.parse(text: '元号甲1年5月1日'),
                           western: Zakuro::Western::Calendar.parse(text: '0446-06-12')
                         ),
